Move the SavePollRating utils under the lib namespace
