+++ lib/cast_vote_msg.js
@@ -5,7 +5,7 @@ const { fetchProp } = require('./utils')
 const { ValidationError, NotFoundError } = require('./errors')
 const CastVote = require('./cast_vote')
 const GetVoteStats = require('./get_vote_stats')
-const SavePollRating = require('../services/save_poll_rating')
+const SavePollRating = require('./save_poll_rating')
 const Reply = require('./reply')
 
 module.exports = function (opts = {}) {
