+++ lib/save_poll_rating.js
@@ -1,6 +1,6 @@
 const Assert = require('assert-plus')
-const { fetchProp } = require('../../lib/utils')
-const { NotFoundError } = require('../../lib/errors')
+const { fetchProp } = require('./utils')
+const { NotFoundError } = require('./errors')
 
 // TODO: Consider renaming SavePollRating to DenormalizePollRating
 //
